repair software amelioration

+++ repair_software.java
@@ -1,5 +1,6 @@
-
 //ouvrir le navigateur par défaut
+import static utils.GameConstants.FPS;
+
 import java.awt.Desktop;
 import java.io.File;
 import java.io.IOException;
@@ -9,7 +10,20 @@ import java.util.List;
 import java.util.Scanner;
 import java.util.concurrent.TimeUnit;
 
-import org.checkerframework.checker.units.qual.s;
+import org.checkerframework.common.value.qual.DoubleVal;
+
+import com.google.common.base.Strings;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.reflect.TypeToken;
+
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.io.FileReader;
+import java.lang.reflect.Type;
 
 public class repair_software {
     public static String directoryPath = "src/main/java/save/";
@@ -28,7 +42,7 @@ public class repair_software {
     }
 
     public static int nombre_sauvegardes() {
-        return files.length - 1;
+        return files.length - 3;
     }
 
     public static void liste_sauvegarde() {
@@ -36,7 +50,7 @@ public class repair_software {
         System.out.println("            Liste des sauvegardes (" + nombre_sauvegardes() + ")");
         if (files != null) {
             for (File file : files) {
-                if (file.getName().endsWith(".json")) {
+                if (file.getName().endsWith(".json") && !file.getName().equals("lastSave.json")){
                     // Validez le fichier de sauvegarde
                     if (isValidSave(file)) {
                         System.out.println("    -" + file.getName() + " ; etat: VALIDE");
@@ -87,11 +101,10 @@ public class repair_software {
 
     public static void gestion_corrompu() {
         System.out.println("voulais-vous supprimer les sauvegardes corrompues ? (oui/non) ");
-        Scanner sc = new Scanner(System.in);
         String reponse = sc.nextLine();
         if (reponse.equals("oui")) {
             for (File file : files) {
-                if (file.getName().endsWith(".json")) {
+                if (file.getName().endsWith(".json") && !file.getName().equals("lastSave.json")) {
                     if (!isValidSave(file)) {
                         file.delete();
                     }
@@ -104,7 +117,6 @@ public class repair_software {
 
     public static void gestion_sauvegarde() {
         System.out.println("voulais-vous supprimer toutes les sauvegardes ? (oui/non) ");
-        Scanner sc = new Scanner(System.in);
         String reponse = sc.nextLine();
         if (reponse.equals("oui")) {
             for (File file : files) {
@@ -118,7 +130,19 @@ public class repair_software {
     }
 
     public static void reparation_automatique() {
-        //reparation automatique
+        System.out.println("voulais-vous réparer toutes les sauvegardes ? (oui/non) ");
+        String reponse = sc.nextLine();
+        if (reponse.equals("oui")) {
+        for (File file : files) {
+            if (file.getName().endsWith(".json") && !file.getName().equals("lastSave.json")) {
+                if (!isValidSave(file)) {
+                    //TODO: Réparer le fichier
+                }
+            }
+        }
+        System.out.println("Réparation en cours...");
+        attendre(120);
+    }
     }
 
     public static void reparation_manuelle() {
@@ -141,13 +165,39 @@ public class repair_software {
         }
     }
 
+    public static void ouvrire_fichier() {
+        File fichier = new File("src/main/java/save");
+        try {
+            Desktop.getDesktop().open(fichier);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+    
+    // Lire les données & obtenir la map de données
+    public static Map<String, Object> getDonnees(File fichier) {
+        Gson gson = new Gson();
+        Map<String, Object> donnees = new HashMap<>();
+        try (FileReader reader = new FileReader(fichier)) {
+            Type type = new TypeToken<Map<String, Object>>() {}.getType();
+            donnees = gson.fromJson(reader, type);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        return donnees;
+    }
+
+
+
+
     public static void action() {
         System.out.println("            que voulez-vous faire ? ");
         System.out.println("#1# suprimer toutes les sauvegardes");
         System.out.println("#2# supprimer les sauvegardes corrompues");
         System.out.println("#3# reparation automatique de la sauvegarde");
         System.out.println("#4# reparation manuelle de la sauvegarde");
-        System.out.println("#5# quitter");
+        System.out.println("#5# ouvrir le fichier de sauvegarde");
+        System.out.println("#6# quitter");
         System.out.println("***********************************************");
         System.out.print("votre choix : ");
         String reponse = sc.nextLine();
@@ -165,6 +215,9 @@ public class repair_software {
                 reparation_manuelle();
                 break;
             case "5":
+                ouvrire_fichier();
+                break;
+            case "6":
                 System.out.println("Fin du programme");
                 fini = true;
                 break;
@@ -175,15 +228,78 @@ public class repair_software {
 
     // Méthode pour valider la sauvegarde
     private static boolean isValidSave(File file) {
-        try {
-            //contenu du fichier de sauvegarde
-            List<String> lines = Files.readAllLines(file.toPath());
-            // Implémentez la logique de validation 
-            return true;
-        } catch (IOException e) {
-            System.err.println("    (SUPPRIMER)");
-            return false; //comme corrompue en cas d'erreur de lecture
-        }
+        Map<String, Object> options = getDonnees(file);
+            if(!(options.get("FPS") instanceof Boolean)) {
+                return false;
+            }
+            if (!(options.get("PATH") instanceof Boolean)) {
+                return false;
+            }
+            if (!(options.get("PARTICLES") instanceof Boolean)) {
+                return false;
+            }
+            if (!(options.get("SOUND") instanceof Number)) {
+                return false;
+            }
+            if (!(options.get("MUSIC") instanceof Number)) {
+                return false;
+            }
+            if (!(options.get("LEFT") instanceof String)) {
+                return false;
+            }
+            if (!(options.get("RIGHT") instanceof String)) {
+                return false;
+            }
+            if (!(options.get("SPACE") instanceof String)) {
+                return false;
+            }
+            if (!(options.get("CSS").equals("DARK") || options.get("CSS").equals("LIGHT") || 
+            options.get("CSS").equals("PINK") || options.get("CSS").equals("CLASSIC") || 
+            options.get("CSS").equals("BLACK") || options.get("CSS").equals("ACHROMATOPSIE") ||
+            options.get("CSS").equals("DEUTERANOPIE") || options.get("CSS").equals("PROTANOPIE") ||
+            options.get("CSS").equals("TRITANOPIE"))){
+                return false;
+            }
+            if(!(options.get("TEXTURE") instanceof String)) {
+                return false;
+            }
+            if(!(options.get("EXP_LEVEL") instanceof Number)) { 
+                return false;
+            }else if(((Number) options.get("EXP_LEVEL")).doubleValue() < 0 ){
+                return false;
+            }
+            if (!(options.get("PROGRESS") instanceof Map)) {
+                return false;
+            }
+            
+            Map<String, Object> progress = (Map<String, Object>) options.get("PROGRESS");
+            
+            if (!progress.containsKey("stages") || !(progress.get("stages") instanceof List)) {
+                return false;
+            }
+            
+            List<Map<String, Object>> stages = (List<Map<String, Object>>) progress.get("stages");
+
+            if( stages.size() != 27){
+                return false;
+            }
+            
+            for (Map<String, Object> stage : stages) {
+                if (!(stage.containsKey("difficulty") && stage.containsKey("unlockLevel") &&
+                      stage.containsKey("completed") && stage.containsKey("maxScore") &&
+                      stage.containsKey("customGame"))) {
+                    return false;
+                }
+            
+                if (!(stage.get("difficulty") instanceof Number && stage.get("unlockLevel") instanceof Number &&
+                      stage.get("completed") instanceof Boolean && stage.get("maxScore") instanceof Number &&
+                      stage.get("customGame") instanceof Boolean)) {
+                    return false;
+                }
+            }
+
+
+        return true;
     }
 
     public static void main(String[] args) {
